+++ libs/cardano-ledger-api/testlib/Test/Cardano/Ledger/Api/Arbitrary.hs
@@ -3,6 +3,7 @@
 module Test.Cardano.Ledger.Api.Arbitrary () where
 
 import Cardano.Ledger.Api.State.Query
+import Cardano.Ledger.Conway.Governance (DefaultVote (..))
 import Generic.Random (genericArbitraryU)
 import Test.Cardano.Ledger.Common
 import Test.Cardano.Ledger.Dijkstra.Arbitrary ()
@@ -10,14 +11,14 @@ import Test.Cardano.Ledger.Dijkstra.Arbitrary ()
 instance Arbitrary MemberStatus where
   arbitrary = arbitraryBoundedEnum
 
-instance Arbitrary QueryPoolStateResult where
-  arbitrary = QueryPoolStateResult <$> arbitrary <*> arbitrary <*> arbitrary <*> arbitrary
+instance Arbitrary QueryResultPoolState where
+  arbitrary = QueryResultPoolState <$> arbitrary <*> arbitrary <*> arbitrary <*> arbitrary
 
-instance Arbitrary StakeSnapshot where
+instance Arbitrary QueryResultStakeSnapshot where
   arbitrary = genericArbitraryU
   shrink = genericShrink
 
-instance Arbitrary StakeSnapshots where
+instance Arbitrary QueryResultStakeSnapshots where
   arbitrary = genericArbitraryU
   shrink = genericShrink
 
@@ -57,4 +58,7 @@ instance Arbitrary QueryResultDRepState where
   arbitrary = genericArbitraryU
 
 instance Arbitrary QueryResultDRepStates where
+  arbitrary = genericArbitraryU
+
+instance Arbitrary QueryResultDelegsAndRewards where
   arbitrary = genericArbitraryU